fix count
