+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core   】     增加UserPassAuthenticator
 
 ### 🐞Bug修复
+* 【db     】     修复Oracle下别名错误造成的SQL语法啊错误（issue#I3VTQW@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
