+++ hutool-db/src/main/java/cn/hutool/db/SqlConnRunner.java
@@ -1,14 +1,11 @@
 package cn.hutool.db;
 
 import cn.hutool.core.collection.CollUtil;
-import cn.hutool.core.lang.Assert;
 import cn.hutool.core.map.MapUtil;
-import cn.hutool.core.util.StrUtil;
 import cn.hutool.db.dialect.Dialect;
 import cn.hutool.db.dialect.DialectFactory;
 import cn.hutool.db.handler.EntityListHandler;
 import cn.hutool.db.handler.HandleHelper;
-import cn.hutool.db.handler.NumberHandler;
 import cn.hutool.db.handler.PageResultHandler;
 import cn.hutool.db.handler.RsHandler;
 import cn.hutool.db.sql.Condition.LikeType;
@@ -283,19 +280,6 @@ public class SqlConnRunner extends DialectRunner {
 	/**
 	 * 获取查询结果总数，生成类似于 SELECT count(1) from (sql) as _count
 	 *
-	 * @param conn       数据库连接对象
-	 * @param sqlBuilder SQL构建器，包括SQL和参数
-	 * @return 结果数
-	 * @throws SQLException SQL异常
-	 * @since 5.7.0
-	 */
-	public long count(Connection conn, SqlBuilder sqlBuilder) throws SQLException {
-		return count(conn, sqlBuilder.build(), sqlBuilder.getParamValueArray());
-	}
-
-	/**
-	 * 获取查询结果总数，生成类似于 SELECT count(1) from (sql) as _count
-	 *
 	 * @param conn      数据库连接对象
 	 * @param selectSql 查询语句
 	 * @param params    查询参数
@@ -304,19 +288,7 @@ public class SqlConnRunner extends DialectRunner {
 	 * @since 5.6.6
 	 */
 	public long count(Connection conn, CharSequence selectSql, Object... params) throws SQLException {
-		Assert.notBlank(selectSql, "Select SQL must be not blank!");
-		final int orderByIndex = StrUtil.indexOfIgnoreCase(selectSql, " order by");
-		if (orderByIndex > 0) {
-			selectSql = StrUtil.subPre(selectSql, orderByIndex);
-		}
-
-		SqlBuilder sqlBuilder = SqlBuilder.of(selectSql)
-				.insertPreFragment("SELECT count(1) from(")
-				// issue#I3IJ8X@Gitee，在子查询时需设置单独别名，此处为了防止和用户的表名冲突，使用自定义的较长别名
-				.append(") as _hutool_alias_count_")
-				// 添加参数
-				.addParams(params);
-		return page(conn, sqlBuilder, null, new NumberHandler()).intValue();
+		return count(conn, SqlBuilder.of(selectSql).addParams(params));
 	}
 
 	/**
